Add files via upload
The Microprocessor Design and Interfacing course delves into the world of designing and connecting microprocessors. We explore the architecture, programming, and practical applications of these essential components. Students learn how to program microprocessors, interface them with various devices, and tackle real-world challenges. The course blends theory with hands-on experience, equipping students with the skills to design and connect microprocessors effectively. From understanding assembly language to implementing interfacing techniques, this course provides a comprehensive foundation for anyone venturing into the dynamic realm of microprocessor design and interfacing.